Time: 5 ms (92.65%), Space: 10.9 MB (64.39%) - LeetHub

+++ 0080-remove-duplicates-from-sorted-array-ii/0080-remove-duplicates-from-sorted-array-ii.c
@@ -1,7 +1,7 @@
 int removeDuplicates(int* nums, int numsSize) {
     bool couple_on = false;
     short counter = 0, k = 1;
-    for (short i = 1; i < (short)numsSize; i++) 
+    for (short i = 1; i < numsSize; i++) 
     {
         if (nums[i] == nums[i - 1] && couple_on ) //dup num
         {
